Add a new TodoList component
